[sys-libs/glibc] Provisional glibc-2.12.1
Provide provisional sys-libs/glibc-2.12.1 ebuilds.

The glibc-2.12.1-r3 is a baseline commit without any Enterprise Linux
patches applied, and the glibc-2.12.1-r11 version is patched with
the Enterprise Linux patch set.

This ebuild is going to the provisional overlay for two primary reasons.
(1)  This is somewhat of an experimental version of glibc.
     The base version for the Enterprise Linux official glibc is 2.12.
     However, this specific version was never a discrete release.
     So, this ebuild is based on the 2.12.1 release, with patches from
     the Enterprise Linux version.
(2)  The patch set is too big, and needs it's own archive.

Signed-off-by: Mitch Harder <mitch.harder@sabayonlinux.org>