[sys-apps/openrc] Add tweaked version of sys-apps/openrc
The sys-apps/openrc-0.11.8 in Gentoo portage has a dependency
on -tinfo for ncurses.  The ncurses package in lts6 has not
implemented a tinfo flag, leaving me in a predicament as to
how to proceed.

Apparently, when building ncurses with +tinfo, you separate
out a tinfo library in order to support some multimedia package.

This is outside the scope of what I intend to provide in the
lts6 version of ncurses anyways.  I'm not sure if there is a
simpler way to do this, but I'm copying the Gentoo version
of the openrc ebuild without the ncurses(-tinfo) dependency.

Signed-off-by: Mitch Harder <mitch.harder@sabayonlinux.org>